requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.7.1
psycopg2-binary==2.9.1
gunicorn==20.1.0
pytz==2021.1
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
aiohttp==3.9.3
